Moved NavItems to NavItems directory
